pyproject.toml
[tool.poetry]
name = "sparkle"
version = "0.1.0"
description = ""
authors = ["Ryana May Que <ryanamayque@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.113.0"}
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
websocket-client = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
